Handler + added user input
