--- GroceryStoreProject/src/main/resources/Prices.cvs
@@ -1,12 +1,0 @@
-Iceberg Lettuce, $1.99, Not Organic
-Romaine Lettuce, $1.99, Not Organic
-Cucumber, $0.79, Not Organic
-American Cucumber, $2.33, Not Organic
-Honeycrisp Apple, $1.20, Not Organic
-Pink Lady Apple, $0.80, Not Organic
-Hass Avocado, $2.50, Organic
-Medium Hass Avocado, $1.00, Not Organic
-Green Onion, $0.89, Organic
-Yellow Onion, $1.43, Organic
-Yellow Onion, $1.03, Not Organic
-Blood Orange, $2.99, Not Organic
